# specs/001-water-plant-discovery/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FlowGrid Water Plant Discovery API
  description: |
    RESTful API for FlowGrid Water Plant Discovery Platform V1.

    ## Authentication
    - Consumer App: JWT Bearer token (username/password login)
    - Owner App: JWT Bearer token (phone/password login)
    - Admin Dashboard: JWT Bearer token (username/password login)

    ## Rate Limiting
    - Public endpoints: 100 requests/minute per IP
    - Authenticated endpoints: 500 requests/minute per user

    ## Response Format
    All responses follow the standard envelope:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "timestamp": "...", "requestId": "..." }
    }
    ```

    Error responses:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message",
        "details": { ... }
      }
    }
    ```
  version: 1.0.0
  contact:
    name: FlowGrid API Support
    email: api@flowgrid.io
servers:
  - url: https://api.flowgrid.io/v1
    description: Production server
  - url: https://staging-api.flowgrid.io/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
...
  /owner/verification:
    get:
      tags:
        - Owner Documents
      summary: Get verification status
      description: Get current verification request status and history
      operationId: getVerificationStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Verification status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Owner Documents
      summary: Submit verification documents
      description: Submit documents for verification (government ID and business registration required)
      operationId: submitVerification
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - governmentId
                - businessRegistration
              properties:
                governmentId:
                  type: string
                  format: binary
                  description: Government ID (Aadhaar/PAN) - JPG, PNG, or PDF
                businessRegistration:
                  type: string
                  format: binary
                  description: Business registration certificate - JPG, PNG, or PDF
                fssaiLicense:
                  type: string
                  format: binary
                  description: FSSAI license (optional) - JPG, PNG, or PDF
      responses:
        '201':
          description: Documents submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationStatusResponse'
        '400':
          description: Already has pending verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ============================================================================
  # OWNER NOTIFICATIONS
  # ============================================================================
...
  /admin/verifications/{requestId}:
    get:
      tags:
        - Admin Verification
      summary: Get verification details
      description: Get full details of a verification request including documents
      operationId: getVerificationDetails
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Verification details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVerificationDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
...
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'UNAUTHORIZED'
              message: 'Invalid or expired token'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'FORBIDDEN'
              message: 'You do not have permission to perform this action'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'NOT_FOUND'
              message: 'Resource not found'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Validation failed'
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    OwnerBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
...
    PlantBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        verificationStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        isOpen:
          type: boolean
...
    VerificationStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            status:
              type: string
              enum: [not_submitted, pending, approved, rejected]
            currentRequest:
              type: object
              nullable: true
              properties:
                id:
                  type: string
                  format: uuid
                status:
                  type: string
                  enum: [pending, approved, rejected]
                submittedAt:
                  type: string
                  format: date-time
                decidedAt:
                  type: string
                  format: date-time
                  nullable: true
                rejectionReason:
                  type: string
                  nullable: true
...
    AdminVerificationDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            request:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                status:
                  type: string
                  enum: [pending, approved, rejected]
                submittedAt:
                  type: string
                  format: date-time
                decidedAt:
                  type: string
                  format: date-time
                  nullable: true
                rejectionReason:
                  type: string
                  nullable: true
                documents:
                  type: object
                  properties:
                    governmentIdUrl:
                      type: string
                      format: uri
                    businessRegistrationUrl:
                      type: string
                      format: uri
                    fssaiLicenseUrl:
                      type: string
                      format: uri
                      nullable: true
            owner:
              $ref: '#/components/schemas/OwnerBasic'
            plant:
              $ref: '#/components/schemas/PlantBasic'

    # ========== Admin ==========
...